+++ R/tm_a_regression.R
@@ -14,6 +14,7 @@
 #' @param size optional, (`numeric`) If scalar then the plot point sizes will have a fixed size
 #'   If a slider should be presented to adjust the plot point sizes dynamically then it can be a
 #'   vector of length three with `c(value, min, max)`.
+#' @param default_outlier_label optional, (`character`) The default column selected to label outliers.
 #' @param default_plot_type optional, (`numeric`) Defaults to Response vs Regressor.
 #'
 #' 1. Response vs Regressor
@@ -79,7 +80,8 @@ tm_a_regression <- function(label = "Regression Analysis",
                             ggtheme = gg_themes,
                             pre_output = NULL,
                             post_output = NULL,
-                            default_plot_type = 1) {
+                            default_plot_type = 1,
+                            default_outlier_label = "USUBJID") {
   if (!is_class_list("data_extract_spec")(regressor)) {
     regressor <- list(regressor)
   }
@@ -100,7 +102,8 @@ tm_a_regression <- function(label = "Regression Analysis",
       ),
     is_class_list("data_extract_spec")(regressor),
     # No check necessary for regressor and response, as checked in data_extract_input
-    is_character_single(ggtheme)
+    is_character_single(ggtheme),
+    is_character_single(default_outlier_label)
     )
   check_slider_input(plot_height, allow_null = FALSE)
   check_slider_input(plot_width)
@@ -118,7 +121,9 @@ tm_a_regression <- function(label = "Regression Analysis",
     server = srv_a_regression,
     ui = ui_a_regression,
     ui_args = args,
-    server_args = c(data_extract_list, list(plot_height = plot_height, plot_width = plot_width)),
+    server_args = c(
+      data_extract_list,
+      list(plot_height = plot_height, plot_width = plot_width, default_outlier_label = default_outlier_label)),
     filters = get_extract_datanames(data_extract_list)
   )
 }
@@ -164,6 +169,21 @@ ui_a_regression <- function(id, ...) {
         choices = plot_choices,
         selected = plot_choices[args$default_plot_type]
       ),
+      checkboxInput(ns("show_outlier"), label = "Display outlier labels", value = TRUE),
+      shinyjs::hidden(optionalSliderInput(
+        ns("outlier"),
+        div(
+          "Outlier definition:",
+          title = paste(
+            "Use the slider to choose the cut-off value to define outliers. Points with a Cook's distance greater than",
+            "the value on the slider times the mean of the Cook's distance of the dataset will have labels."),
+          icon("info-circle")
+        ), min = 1, max = 10, value = 9, ticks = FALSE, step = .1)),
+      shinyjs::hidden(optionalSelectInput(
+        ns("label_var"),
+        multiple = FALSE,
+        label = "Outlier label"
+      )),
       panel_group(
         panel_item(
           title = "Plot settings",
@@ -187,6 +207,7 @@ ui_a_regression <- function(id, ...) {
 
 
 #' @importFrom magrittr %>%
+#' @importFrom dplyr if_else
 #' @importFrom methods is substituteDirect
 #' @importFrom stats as.formula
 #' @importFrom stats lowess
@@ -198,13 +219,20 @@ srv_a_regression <- function(input,
                              response,
                              regressor,
                              plot_height,
-                             plot_width) {
+                             plot_width,
+                             default_outlier_label) {
   init_chunks()
 
   merged_data <- data_merge_module(
     datasets = datasets,
     data_extract = list(response, regressor),
     input_id = c("response", "regressor")
+  )
+
+  regression_var <- reactive(
+    list(
+      response = as.vector(merged_data()$columns_source$response),
+      regressor = as.vector(merged_data()$columns_source$regressor))
   )
 
   # sets chunk object and populates it with data merge call and fit expression
@@ -218,37 +246,47 @@ srv_a_regression <- function(input,
     ANL <- chunks_get_var("ANL", chunks = chunks_stack) # nolint
     validate_has_data(ANL, 10)
 
-    response_var <- as.vector(merged_data()$columns_source$response)
-    regressor_var <- as.vector(merged_data()$columns_source$regressor)
+    opts <- variable_choices(ANL)
+    selected <- if (!is.null(isolate(input$label_var)) && isolate(input$label_var) %in% as.character(opts)) {
+      isolate(input$label_var)
+      } else {
+        if_empty(opts[as.character(opts) == default_outlier_label], opts[[1]])
+      }
+    updateOptionalSelectInput(
+      session = session,
+      inputId = "label_var",
+      choices = opts,
+      selected = selected)
 
     # validation
     validate(
       need(
-        length(regressor_var) > 0,
+        length(regression_var()$regressor) > 0,
         "At least one regressor should be selected."
         )
       )
     validate(
       need(
-        length(response_var) == 1,
+        length(regression_var()$response) == 1,
         "Response variable should be of length one."
         )
       )
-    validate(need(is.numeric(ANL[response_var][[1]]), "Response variable should be numeric."))
+    validate(need(is.numeric(ANL[regression_var()$response][[1]]), "Response variable should be numeric."))
     validate(
       need(
-        input$plot_type != "Response vs Regressor" || length(regressor_var) == 1,
+        input$plot_type != "Response vs Regressor" || length(regression_var()$regressor) == 1,
         "Response vs Regressor is only provided for exactly one regressor"
         )
       )
 
-    validate_has_data(ANL[, c(response_var, regressor_var)], 10, complete = TRUE, allow_inf = FALSE)
+    validate_has_data(
+      ANL[, c(regression_var()$response, regression_var()$regressor)], 10, complete = TRUE, allow_inf = FALSE)
 
     form <- stats::as.formula(
       paste(
-        response_var,
+        regression_var()$response,
         paste(
-          regressor_var,
+          regression_var()$regressor,
           collapse = " + "
         ),
         sep = " ~ "
@@ -273,7 +311,48 @@ srv_a_regression <- function(input,
 
     chunks_safe_eval(chunks = chunks_stack)
 
+    data <- fortify(lm(form, data = ANL))
+
+    cooksd <- data$.cooksd[!is.nan(data$.cooksd)]
+    max_outlier <- max(ceiling(max(cooksd) / mean(cooksd)), 2)
+
+    cur_outlier <- isolate(input$outlier)
+    updateSliderInput(
+      session = session,
+      inputId = "outlier",
+      min = 1,
+      max = max_outlier,
+      value = if (cur_outlier < max_outlier) cur_outlier else max_outlier * .9)
+
     chunks_stack
+  })
+
+  label_col <- reactive({
+    validate(need(input$label_var, "`Display outlier labels` field is checked but `Outlier label` field is empty"))
+    bquote(
+      dplyr::if_else(
+        data$.cooksd > .(input$outlier) * mean(data$.cooksd, na.rm = TRUE),
+        as.character(ANL[[.(input$label_var)]]),
+        "") %>% dplyr::if_else(is.na(.), "cooksd == NaN", .)
+    )
+  })
+
+  outlier_label <- reactive({
+    bquote(geom_text(
+      label = .(label_col()),
+      hjust = 0,
+      vjust = 1,
+      color = "red"))
+  })
+
+  observeEvent(input$show_outlier, {
+    if (input$show_outlier) {
+      shinyjs::show("outlier")
+      shinyjs::show("label_var")
+    } else {
+      shinyjs::hide("outlier")
+      shinyjs::hide("label_var")
+    }
   })
 
 
@@ -282,6 +361,7 @@ srv_a_regression <- function(input,
     size <- input$size # nolint
     ggtheme <- input$ggtheme # nolint
     input_type <- input$plot_type
+    show_outlier <- input$show_outlier
 
     chunks_reset()
     chunks_push_chunks(fit())
@@ -290,42 +370,56 @@ srv_a_regression <- function(input,
 
     plot_type_0 <- function() {
       fit <- chunks_get_var("fit") # chunk already evaluated
-      response_var <- as.vector(merged_data()$columns_source$response) # nolint
-      regressor_var <- as.vector(merged_data()$columns_source$regressor) # nolint
       ANL <- chunks_get_var("ANL") # nolint
 
       stopifnot(ncol(fit$model) == 2)
 
-      if (!is.factor(ANL[[regressor_var]])) {
+      if (!is.factor(ANL[[regression_var()$regressor]])) {
         shinyjs::show("size")
         shinyjs::show("alpha")
-        chunks_push(id = "plot_0_a", expression = bquote({
-          gg <- ggplot(
-            fit$model[, 2:1],
-            aes_string(.(regressor_var), .(response_var))
+        plot <- bquote(ggplot(
+          fit$model[, 2:1],
+          aes_string(.(regression_var()$regressor), .(regression_var()$response))
           ) +
-            geom_point(size = .(size), alpha = .(alpha)) +
-            stat_smooth(
-              method = "lm",
-              formula = y ~ x,
-              se = FALSE
-            )
+          geom_point(size = .(size), alpha = .(alpha)) +
+          stat_smooth(
+            method = "lm",
+            formula = y ~ x,
+            se = FALSE
+          )
+        )
+        if (show_outlier) {
+          plot <- bquote(.(plot) + .(outlier_label()))
+        }
+        chunks_push(id = "plot_0_a", expression = bquote({
+          class(fit$residuals) <- NULL
+          data <- fortify(fit)
+          gg <- .(plot)
         }))
       } else {
         shinyjs::hide("size")
         shinyjs::hide("alpha")
-        chunks_push(id = "plot_0_a", expression = bquote({
-          gg <- ggplot(
+        plot <- bquote(
+          ggplot(
             fit$model[, 2:1],
-            aes_string(.(regressor_var), .(response_var))
-          ) + geom_boxplot()
+            aes_string(.(regression_var()$regressor), .(regression_var()$response))
+          ) +
+          geom_boxplot()
+        )
+        if (show_outlier) {
+          plot <- bquote(.(plot) + .(outlier_label()))
+        }
+        chunks_push(id = "plot_0_a", expression = bquote({
+          class(fit$residuals) <- NULL
+          data <- fortify(fit)
+          gg <- .(plot)
         }))
       }
 
       chunks_push(id = "plot_0_b", expression = bquote({
         g_final <- gg +
-          xlab(.(varname_w_label(regressor_var, ANL))) +
-          ylab(.(varname_w_label(response_var, ANL))) +
+          xlab(.(varname_w_label(regression_var()$regressor, ANL))) +
+          ylab(.(varname_w_label(regression_var()$response, ANL))) +
           ggtitle("Response vs Regressor") +
           .(call(paste0("theme_", ggtheme)))
         print(g_final)
@@ -351,16 +445,17 @@ srv_a_regression <- function(input,
     plot_type_1 <- function() {
       shinyjs::show("size")
       shinyjs::show("alpha")
+      plot <- bquote(
+        ggplot(data = data, aes(.fitted, .resid)) +
+        geom_point(size = .(size), alpha = .(alpha)) +
+        geom_hline(yintercept = 0, linetype = "dashed", size = 1) +
+        geom_line(data = smoothy, mapping = smoothy_aes))
+      if (show_outlier) {
+        plot <- bquote(.(plot) + .(outlier_label()))
+      }
       chunks_push(id = "plot_1", expression = bquote({
         smoothy <- smooth(data$.fitted, data$.resid)
-        g_base <- ggplot(data = data, aes(.fitted, .resid)) +
-          geom_point(size = .(size), alpha = .(alpha)) +
-          geom_hline(
-            yintercept = 0,
-            linetype = "dashed",
-            size = 1
-          ) +
-          geom_line(data = smoothy, mapping = smoothy_aes)
+        g_base <- .(plot)
         g_final <- g_base + labs(
           x = paste0("Fitted values\nlm(", reg_form, ")"),
           y = "Residuals",
@@ -373,10 +468,25 @@ srv_a_regression <- function(input,
     plot_type_2 <- function() {
       shinyjs::show("size")
       shinyjs::show("alpha")
-      chunks_push(id = "plot_2", expression = bquote({
-        g_base <- ggplot(data = data) +
-          stat_qq(aes(sample = .stdresid), size = .(size), alpha = .(alpha)) +
+      plot <- bquote(
+        ggplot(data = data, aes(sample = .stdresid)) +
+          stat_qq(size = .(size), alpha = .(alpha)) +
           geom_abline(linetype = "dashed")
+      )
+      if (show_outlier) {
+        plot <- bquote(
+          .(plot) +
+            stat_qq(
+              geom = "text",
+              label = .(label_col()) %>% `[`(. != "cooksd == NaN"),
+              hjust = 0,
+              vjust = 1,
+              color = "red"
+            )
+          )
+      }
+      chunks_push(id = "plot_2", expression = bquote({
+        g_base <- .(plot)
         g_final <- g_base + labs(
           x = paste0("Theoretical Quantiles\nlm(", reg_form, ")"),
           y = "Standardized residuals",
@@ -390,11 +500,17 @@ srv_a_regression <- function(input,
     plot_type_3 <- function() {
       shinyjs::show("size")
       shinyjs::show("alpha")
-      chunks_push(id = "plot_3", expression = bquote({
-        smoothy <- smooth(data$.fitted, sqrt(abs(data$.stdresid)))
-        g_base <- ggplot(data = data, aes(.fitted, sqrt(abs(.stdresid)))) +
+      plot <- bquote(
+        ggplot(data = data, aes(.fitted, sqrt(abs(.stdresid)))) +
           geom_point(size = .(size), alpha = .(alpha)) +
           geom_line(data = smoothy, mapping = smoothy_aes)
+      )
+      if (show_outlier) {
+        plot <- bquote(.(plot) + .(outlier_label()))
+      }
+      chunks_push(id = "plot_3", expression = bquote({
+        smoothy <- smooth(data$.fitted, sqrt(abs(data$.stdresid)))
+        g_base <- .(plot)
         g_final <- g_base + labs(
           x = paste0("Fitted values\nlm(", reg_form, ")"),
           y = expression(sqrt(abs(`Standardized residuals`))),
@@ -408,9 +524,35 @@ srv_a_regression <- function(input,
     plot_type_4 <- function() {
       shinyjs::hide("size")
       shinyjs::show("alpha")
-      chunks_push(id = "plot_4", expression = bquote({
-        g_base <- ggplot(data = data, aes(seq_along(.cooksd), .cooksd)) +
+      plot <- bquote(
+        ggplot(data = data, aes(seq_along(.cooksd), .cooksd)) +
           geom_col(alpha = .(alpha))
+      )
+      if (show_outlier) {
+        plot <- bquote(
+          .(plot) +
+            geom_hline(
+              yintercept = c(
+                .(input$outlier) * mean(data$.cooksd, na.rm = TRUE),
+                mean(data$.cooksd, na.rm = TRUE)),
+              color = "red",
+              linetype = "dashed") +
+            geom_text(
+              aes(
+                x = 0,
+                y = mean(data$.cooksd, na.rm = TRUE),
+                label = paste("mu", "=", round(mean(data$.cooksd, na.rm = TRUE), 4)),
+                vjust = -1,
+                hjust = 0,
+                color = "red",
+                angle = 90),
+              parse = TRUE,
+              show.legend = FALSE) +
+            .(outlier_label())
+          )
+      }
+      chunks_push(id = "plot_4", expression = bquote({
+        g_base <- .(plot)
         g_final <- g_base + labs(
           x = paste0("Obs. number\nlm(", reg_form, ")"),
           y = "Cook's distance",
@@ -425,9 +567,8 @@ srv_a_regression <- function(input,
     plot_type_5 <- function() {
       shinyjs::show("size")
       shinyjs::show("alpha")
-      chunks_push(id = "plot_5", expression = bquote({
-        smoothy <- smooth(data$.hat, data$.stdresid)
-        g_base <- ggplot(data = data, aes(.hat, .stdresid)) +
+      plot <- bquote(
+        ggplot(data = data, aes(.hat, .stdresid)) +
           geom_vline(
             size = 1,
             colour = "black",
@@ -442,6 +583,13 @@ srv_a_regression <- function(input,
           ) +
           geom_point(size = .(size), alpha = .(alpha)) +
           geom_line(data = smoothy, mapping = smoothy_aes)
+      )
+      if (show_outlier) {
+        plot <- bquote(.(plot) + .(outlier_label()))
+      }
+      chunks_push(id = "plot_5", expression = bquote({
+        smoothy <- smooth(data$.hat, data$.stdresid)
+        g_base <- .(plot)
         g_final <- g_base + labs(
           x = paste0("Standardized residuals\nlm(", reg_form, ")"),
           y = "Leverage",
@@ -456,9 +604,8 @@ srv_a_regression <- function(input,
     plot_type_6 <- function() {
       shinyjs::show("size")
       shinyjs::show("alpha")
-      chunks_push(id = "plot_6", expression = bquote({
-        smoothy <- smooth(data$.hat, data$.cooksd)
-        g_base <- ggplot(data = data, aes(.hat, .cooksd)) +
+      plot <- bquote(
+        ggplot(data = data, aes(.hat, .cooksd)) +
           geom_vline(xintercept = 0, colour = NA) +
           geom_abline(
             slope = seq(0, 3, by = 0.5),
@@ -468,6 +615,13 @@ srv_a_regression <- function(input,
           ) +
           geom_line(data = smoothy, mapping = smoothy_aes) +
           geom_point(size = .(size), alpha = .(alpha))
+      )
+      if (show_outlier) {
+        plot <- bquote(.(plot) + .(outlier_label()))
+      }
+      chunks_push(id = "plot_6", expression = bquote({
+        smoothy <- smooth(data$.hat, data$.cooksd)
+        g_base <- .(plot)
         g_final <- g_base + labs(
           x = paste0("Leverage\nlm(", reg_form, ")"),
           y = "Cooks's distance",